Rename Dropdown to index
