--- src/components/Dropdown/Dropdown.js
@@ -1,73 +1,0 @@
-import React from 'react';
-import classNames from 'classnames';
-import PropTypes from 'prop-types';
-
-class Dropdown extends React.Component {
-  render() {
-    const {
-      children,
-      handleClick,
-      icon,
-      isDisabled,
-      isOpen,
-      fullWidth,
-      style,
-      text,
-      width,
-      zIndex,
-    } = this.props;
-
-    const buttonClass = classNames(
-      'oui-button',
-      { [`oui-button--${style}`]: style },
-      { ['oui-button--full soft--left text--left']: fullWidth }
-    );
-
-    const iconClass = classNames(
-      'push-half--left',
-      {
-        ['oui-arrow-inline--down']: icon === 'triangle',
-        ['lego-arrow-inline--down']: icon === 'chevron',
-      }
-    );
-
-    return (
-      <div
-        data-ui-component={ true }
-        ref='dropdown'
-        className='lego-dropdown-group'
-        style={{ width: width, zIndex: zIndex }}>
-        <button
-          className={ buttonClass }
-          disabled={ isDisabled }
-          onClick={ handleClick }
-          style={ { marginBottom: 2 } }>
-          <div className='flex'>
-            <div className='flex--1 truncate'>{ text }</div>
-            <div className='text--right'><span className={ iconClass }></span></div>
-          </div>
-        </button>
-        { isOpen && !isDisabled && children }
-      </div>
-    );
-  }
-};
-
-Dropdown.propTypes = {
-  activator: PropTypes.node.isRequired,
-  children: PropTypes.node.isRequired,
-  fullWidth: PropTypes.bool,
-  handleClick: PropTypes.func,
-  icon: PropTypes.string,
-  isDisabled: PropTypes.bool,
-  isOpen: PropTypes.bool,
-  style: PropTypes.string,
-  text: PropTypes.string,
-  width: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-  ]),
-  zIndex: PropTypes.number,
-};
-
-export default Dropdown;
